+++ src/cli/commands/mc-version-help.ts
@@ -15,7 +15,7 @@ export default (program: CommanderStatic) => {
         log(`\n  Documentation:\n`);
         log(`    the ${magenta("magenta colored commands *")} require service credentials.`);
         log(`    the ${cyan("cyan colored commands ")}require mindconnectlib (agent) credentials`);
-        log(`    the ${blue("blue colored commands #")} use analytical functions of MindSphere @`);
+        log(`    the ${blue("blue colored commands @")} use analytical functions of MindSphere`);
         log(`    the ${green("green colored commands #")} are used to setup starter projects`);
         log(`    the service credentials should only be used in secure environments for setup tasks`);
         log(`    Full documentation: ${cyan("https://mindsphere.github.io")}\n`);